Synchronize with 3.09. Works with a patched version of OCaml 3.09
git-svn-id: http://caml.inria.fr/svn/ocaml/branches/cducetrunk@7218 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/win32unix/unixsupport.h
@@ -46,7 +46,7 @@ extern value win_alloc_socket(SOCKET);
 #define NO_CRT_FD (-1)
 #define Nothing ((value) 0)
 
-extern void win32_maperr(unsigned long errcode);
+extern void win32_maperr(DWORD errcode);
 extern void unix_error (int errcode, char * cmdname, value arg);
 extern void uerror (char * cmdname, value arg);
 extern value unix_freeze_buffer (value);
